requirements.txt
Django==4.2.4
django-debug-toolbar==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
Pillow==10.0.0
rest-framework-simplejwt==0.0.2
python-dotenv==1.0.0
psycopg2-binary==2.9.7
